Function properly when grep is set to color
I have grep set to highlight its output using the GREP_COLOR enviroment
variable. Because brew-which-formula shells out to grep, the color
output confuses it.

This pull request passes the --color=never option to grep, which solves
the problem.

However, I'd recommend trying to get away from shelling out to grep if
possible. There'll probably be a lot more headaches like this that are
system-dependent and difficult to track down.

Closes #22.

Signed-off-by: Baptiste Fontaine <batifon@yahoo.fr>